+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   * Make `PayPalCreditFinancing#fromJson()` package-private
   * Make `PayPalCreditFinancingAmount#fromJson()` package-private
   * Make `UnionPayCapabilities#fromJson()` package-private
-  * Make `PaymentMethodClient` package-private
+  * Make `PaymentMethodClient#parsePaymentMethodNonces()` package-private
 
 ## 4.0.0-beta3